requirements.txt
streamlit
pymongo
pymupdf
python-docx
